code/backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
duckdb>=0.10.0
scikit-learn>=1.4.0
nltk>=3.8.1
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
rank_bm25>=0.2.2
